requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.2=py38hecd8cb5_1001
argon2-cffi=20.1.0=py38haf1e3a3_1
async_generator=1.10=py_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
ca-certificates=2020.12.8=hecd8cb5_0
certifi=2020.12.5=py38hecd8cb5_0
cffi=1.14.4=py38h2125817_0
cloudpickle=1.6.0=pypi_0
cycler=0.10.0=py38_0
dbus=1.13.18=h18a8e69_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py38_0
expat=2.2.10=hb1e8313_2
freetype=2.10.4=ha233b18_0
future=0.18.2=pypi_0
gettext=0.19.8.1=hb0f4f8b_2
glib=2.66.1=h9bbe63b_0
gym=0.17.3=pypi_0
icu=58.2=h0a44026_3
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2019.4=233
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.19.0=py38h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.5.1=py_1
jedi=0.17.2=py38hecd8cb5_1
jinja2=2.11.2=py_0
joblib=0.17.0=py_0
jpeg=9b=he5867d9_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py38_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.7.0=py38hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.0=py38h23ab428_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
llvm-openmp=10.0.0=h28b9765_0
lru-dict=1.1.6=pypi_0
lz4-c=1.9.2=h79c402e_3
markupsafe=1.1.1=py38h1de35cc_1
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38h181983e_0
mistune=0.8.4=py38h1de35cc_1001
mkl=2019.4=233
mkl-service=2.3.0=py38hfbe908c_0
mkl_fft=1.2.0=py38hc64f4ea_0
mkl_random=1.1.1=py38h959d312_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.4.3=pyhd3eb1b0_0
ninja=1.10.1=py38h879752b_0
notebook=6.1.4=py38_0
numpy=1.19.2=py38h456fd55_0
numpy-base=1.19.2=py38hcfb5961_0
olefile=0.46=py_0
openssl=1.1.1i=h9ed2024_0
packaging=20.7=pyhd3eb1b0_0
pandas=1.1.3=py38hb1e8313_0
pandoc=2.11=h0dc7051_0
pandocfilters=1.4.3=py38hecd8cb5_1
parso=0.7.0=py_0
pcre=8.44=hb1e8313_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.0.1=py38h5270095_0
pip=20.2.4=py38hecd8cb5_0
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
ptyprocess=0.6.0=pyhd3eb1b0_2
pycparser=2.20=py_2
pyglet=1.5.0=pypi_0
pygments=2.7.2=pyhd3eb1b0_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h655552a_2
pyrsistent=0.17.3=py38haf1e3a3_0
python=3.8.5=h26836e1_1
python-dateutil=2.8.1=py_0
pytorch=1.7.0=py3.8_0
pytz=2020.1=py_0
pyzmq=20.0.0=py38h23ab428_1
qt=5.9.7=h468cd18_1
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=8.0=h1de35cc_0
scikit-learn=0.23.2=py38h959d312_0
scipy=1.5.4=pypi_0
seaborn=0.11.0=py_0
send2trash=1.5.0=py38_0
setuptools=50.3.1=py38hecd8cb5_1
sip=4.19.8=py38h0a44026_0
six=1.15.0=py38hecd8cb5_0
sqlite=3.33.0=hffcf06c_0
terminado=0.9.1=py38_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hb0a8c7a_0
torchvision=0.8.1=py38_cpu
tornado=6.1=py38h9ed2024_0
traitlets=5.0.5=py_0
typedload=2.3=pypi_0
typing_extensions=3.7.4.3=py_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
wheel=0.35.1=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38_0
wiki-dump-parser=2.0.0=pypi_0
wikiextractor=0.1=pypi_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.3=hb1e8313_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
